Azure-Chat-App-API-master/requirements.txt
Flask ~= 2.2.2
msal ~= 1.21.0
azure-identity ~= 1.12.0
pymssql ~= 2.2.7
azure-storage-blob ~= 12.16.0
pandas ~= 2.0.0
pyarrow ~= 11.0.0
flask_cors ~= 3.0.10
pyodbc ~= 4.0.39